refactor(ui): split management application into focused modules
