+++ site/assets/llm-proxy/js/core/backendClient.js
@@ -1,6 +1,6 @@
 // @ts-check
 
-import { MPR_UI } from "../constants.js?v=20260809b123";
+import { MPR_UI } from "../constants.js?v=20260809i217";
 
 const MANAGEMENT_BASE_PATH = "/api/management";
 const HEADER_CONTENT_TYPE = "Content-Type";
