+++ site/assets/llm-proxy/js/startupGuard.js
@@ -1,7 +1,7 @@
 // @ts-check
 
-import { RUNTIME_UI } from "./constants.js?v=20260809b123";
-import { failApplicationStartup } from "./ui/applicationStartup.js?v=20260809b123";
+import { RUNTIME_UI } from "./constants.js?v=20260809i217";
+import { failApplicationStartup } from "./ui/applicationStartup.js?v=20260809i217";
 
 const applicationModule = document.getElementById(RUNTIME_UI.APPLICATION_MODULE_ID);
 if (!applicationModule) {
