+++ site/assets/llm-proxy/js/ui/managementApplication.js
@@ -1,0 +1,61 @@
+// @ts-check
+
+import { APP_INTEGRITY_ERROR } from "../constants.js?v=20260809i217";
+import { createAdminDashboardResponsibility } from "./adminDashboard.js?v=20260809i217";
+import { createAuthenticationLifecycleResponsibility } from "./authenticationLifecycle.js?v=20260809i217";
+import { createClientAccessResponsibility } from "./clientAccess.js?v=20260809i217";
+import { createManagementApplicationPresentationResponsibility } from "./managementApplicationPresentation.js?v=20260809i217";
+import { createManagementApplicationState } from "./managementApplicationState.js?v=20260809i217";
+import { createNotificationsResponsibility } from "./notifications.js?v=20260809i217";
+import { createProfileMutationsResponsibility } from "./profileMutations.js?v=20260809i217";
+import { createProviderCredentialsResponsibility } from "./providerCredentials.js?v=20260809i217";
+import { createProviderEditorResponsibility } from "./providerEditor.js?v=20260809i217";
+import { createProviderSettingsResponsibility } from "./providerSettings.js?v=20260809i217";
+import { createRequestExamplesResponsibility } from "./requestExamples.js?v=20260809i217";
+import { createRoutingDefaultsResponsibility } from "./routingDefaults.js?v=20260809i217";
+import { createSettingsDialogResponsibility } from "./settingsDialog.js?v=20260809i217";
+import { createTenantSettingsResponsibility } from "./tenantSettings.js?v=20260809i217";
+import { createUsageDashboardResponsibility } from "./usageDashboard.js?v=20260809i217";
+
+/**
+ * Compose the authenticated management application from non-overlapping UI responsibilities.
+ *
+ * @returns {object}
+ */
+export function createManagementApplication() {
+  return composeManagementApplication(
+    createManagementApplicationState(),
+    createManagementApplicationPresentationResponsibility(),
+    createAuthenticationLifecycleResponsibility(),
+    createNotificationsResponsibility(),
+    createProfileMutationsResponsibility(),
+    createTenantSettingsResponsibility(),
+    createUsageDashboardResponsibility(),
+    createAdminDashboardResponsibility(),
+    createSettingsDialogResponsibility(),
+    createProviderEditorResponsibility(),
+    createProviderCredentialsResponsibility(),
+    createProviderSettingsResponsibility(),
+    createRoutingDefaultsResponsibility(),
+    createClientAccessResponsibility(),
+    createRequestExamplesResponsibility(),
+  );
+}
+
+/**
+ * @param {...object} responsibilities
+ * @returns {object}
+ */
+function composeManagementApplication(...responsibilities) {
+  const application = {};
+  for (const responsibility of responsibilities) {
+    const descriptors = Object.getOwnPropertyDescriptors(responsibility);
+    for (const propertyName of Object.keys(descriptors)) {
+      if (Object.hasOwn(application, propertyName)) {
+        throw new Error(`${APP_INTEGRITY_ERROR}:duplicate_management_application_property:${propertyName}`);
+      }
+    }
+    Object.defineProperties(application, descriptors);
+  }
+  return application;
+}
