+++ site/assets/llm-proxy/js/ui/managementApplicationState.js
@@ -1,0 +1,188 @@
+// @ts-check
+
+import {
+  AUTH_STATES,
+  COPY,
+  DASHBOARD_VIEWS,
+  DEFAULT_USAGE_INTERVAL,
+  NOTICE_KINDS,
+  NOTICE_SURFACES,
+  USAGE_INTERVALS,
+} from "../constants.js?v=20260809i217";
+import { emptyDefaults } from "../core/managementProfile.js?v=20260809i217";
+import { createProviderEditorSession } from "./providerEditor.js?v=20260809i217";
+import { emptyUsageSummary } from "./usagePresentation.js?v=20260809i217";
+
+const EMPTY_STRING = "";
+
+/**
+ * Create the complete reactive state owned by the authenticated management application.
+ *
+ * @returns {{
+ *   states: { loading: string, authenticated: string, unauthenticated: string, error: string },
+ *   dashboardViews: typeof DASHBOARD_VIEWS,
+ *   copy: typeof COPY,
+ *   authState: import("../types.d.js").AuthState,
+ *   busy: boolean,
+ *   dashboardView: import("../types.d.js").DashboardView,
+ *   usageIntervals: typeof USAGE_INTERVALS,
+ *   selectedUsageInterval: import("../types.d.js").UsageInterval,
+ *   selectedUsageTenantID: string,
+ *   usageLoading: boolean,
+ *   usageLoadVersion: number,
+ *   account: import("../types.d.js").ManagementAccount | null,
+ *   tenants: import("../types.d.js").ManagementTenantSummary[],
+ *   settingsTenantID: string,
+ *   profile: import("../types.d.js").ManagementTenantProfile | null,
+ *   runtimeConfig: import("../types.d.js").FrontendRuntimeConfig | null,
+ *   providers: import("../types.d.js").ProviderProfile[],
+ *   providerEditorSession: import("../types.d.js").ProviderEditorSession,
+ *   providerAutosavePending: boolean,
+ *   providerAutosavePromise: Promise<boolean> | null,
+ *   providerKeyVerificationPending: boolean,
+ *   providerKeyVerificationController: AbortController | null,
+ *   providerKeyVerificationFailure: string,
+ *   routingDefaultsDirty: boolean,
+ *   routingDefaultsEditVersion: number,
+ *   routingProviderSelectionPending: boolean,
+ *   routingProviderSelectionVersion: number,
+ *   routingDefaultsAutosavePending: boolean,
+ *   routingDefaultsAutosavePromise: Promise<boolean> | null,
+ *   profileMutationTail: Promise<void>,
+ *   profileMutationFailureVersion: number,
+ *   clientKeyMutationPromise: Promise<boolean> | null,
+ *   profileApplicationVersion: number,
+ *   providerRemovalConfirmationProviderID: string,
+ *   defaults: import("../types.d.js").TenantDefaults,
+ *   usage: import("../types.d.js").ManagementUsageSummary,
+ *   usageFailuresOpen: boolean,
+ *   usageFailuresLoading: boolean,
+ *   usageFailuresError: string,
+ *   usageFailuresLoadVersion: number,
+ *   usageFailures: Array<import("../types.d.js").ManagementUsageFailure | import("../types.d.js").ManagementAccountUsageFailure>,
+ *   usageFailuresNextCursor: string,
+ *   adminUsers: import("../types.d.js").ManagementAdminUser[],
+ *   appLoadPromise: Promise<void> | null,
+ *   appVersion: number,
+ *   accountRequestController: AbortController | null,
+ *   tenantLifetimeController: AbortController | null,
+ *   tenantRequestController: AbortController | null,
+ *   usageRequestController: AbortController | null,
+ *   usageFailuresRequestController: AbortController | null,
+ *   generatedSecret: string,
+ *   generatedSecretVisible: boolean,
+ *   generatedSecretVersion: number,
+ *   settingsOpen: boolean,
+ *   settingsClosePending: boolean,
+ *   usageExamplesOpen: boolean,
+ *   routingSystemPromptOpen: boolean,
+ *   providerSystemPromptOpen: boolean,
+ *   tenantNameDraft: string,
+ *   tenantRenameDialogOpen: boolean,
+ *   tenantNameDirty: boolean,
+ *   tenantNameError: string,
+ *   clientKeyReplacementConfirmationOpen: boolean,
+ *   clientKeyReplacementPending: boolean,
+ *   createTenantDialogOpen: boolean,
+ *   createTenantName: string,
+ *   createTenantError: string,
+ *   createTenantPending: boolean,
+ *   deleteTenantConfirmationOpen: boolean,
+ *   deleteTenantPending: boolean,
+ *   discardTenantChangesOpen: boolean,
+ *   pendingTenantID: string,
+ *   noticeSurfaces: typeof NOTICE_SURFACES,
+ *   notice: { kind: string, message: string, surface: string },
+ *   noticeDismissTimerID: number | null,
+ *   noticeVersion: number
+ * }}
+ */
+export function createManagementApplicationState() {
+  return {
+    states: {
+      loading: AUTH_STATES.LOADING,
+      authenticated: AUTH_STATES.AUTHENTICATED,
+      unauthenticated: AUTH_STATES.UNAUTHENTICATED,
+      error: AUTH_STATES.ERROR,
+    },
+    dashboardViews: DASHBOARD_VIEWS,
+    copy: COPY,
+    authState: /** @type {import("../types.d.js").AuthState} */ (AUTH_STATES.LOADING),
+    busy: false,
+    dashboardView: /** @type {import("../types.d.js").DashboardView} */ (DASHBOARD_VIEWS.USAGE),
+    usageIntervals: USAGE_INTERVALS,
+    selectedUsageInterval: /** @type {import("../types.d.js").UsageInterval} */ (DEFAULT_USAGE_INTERVAL),
+    selectedUsageTenantID: EMPTY_STRING,
+    usageLoading: false,
+    usageLoadVersion: 0,
+    account: /** @type {import("../types.d.js").ManagementAccount | null} */ (null),
+    tenants: /** @type {import("../types.d.js").ManagementTenantSummary[]} */ ([]),
+    settingsTenantID: EMPTY_STRING,
+    profile: /** @type {import("../types.d.js").ManagementTenantProfile | null} */ (null),
+    runtimeConfig: /** @type {import("../types.d.js").FrontendRuntimeConfig | null} */ (null),
+    providers: /** @type {import("../types.d.js").ProviderProfile[]} */ ([]),
+    providerEditorSession: createProviderEditorSession(EMPTY_STRING, 0),
+    providerAutosavePending: false,
+    providerAutosavePromise: /** @type {Promise<boolean> | null} */ (null),
+    providerKeyVerificationPending: false,
+    providerKeyVerificationController: /** @type {AbortController | null} */ (null),
+    providerKeyVerificationFailure: EMPTY_STRING,
+    routingDefaultsDirty: false,
+    routingDefaultsEditVersion: 0,
+    routingProviderSelectionPending: false,
+    routingProviderSelectionVersion: 0,
+    routingDefaultsAutosavePending: false,
+    routingDefaultsAutosavePromise: /** @type {Promise<boolean> | null} */ (null),
+    profileMutationTail: Promise.resolve(),
+    profileMutationFailureVersion: 0,
+    clientKeyMutationPromise: /** @type {Promise<boolean> | null} */ (null),
+    profileApplicationVersion: 0,
+    providerRemovalConfirmationProviderID: EMPTY_STRING,
+    defaults: emptyDefaults(),
+    usage: emptyUsageSummary(DEFAULT_USAGE_INTERVAL),
+    usageFailuresOpen: false,
+    usageFailuresLoading: false,
+    usageFailuresError: EMPTY_STRING,
+    usageFailuresLoadVersion: 0,
+    usageFailures: /** @type {Array<import("../types.d.js").ManagementUsageFailure | import("../types.d.js").ManagementAccountUsageFailure>} */ ([]),
+    usageFailuresNextCursor: EMPTY_STRING,
+    adminUsers: /** @type {import("../types.d.js").ManagementAdminUser[]} */ ([]),
+    appLoadPromise: /** @type {Promise<void> | null} */ (null),
+    appVersion: 0,
+    accountRequestController: /** @type {AbortController | null} */ (null),
+    tenantLifetimeController: /** @type {AbortController | null} */ (null),
+    tenantRequestController: /** @type {AbortController | null} */ (null),
+    usageRequestController: /** @type {AbortController | null} */ (null),
+    usageFailuresRequestController: /** @type {AbortController | null} */ (null),
+    generatedSecret: EMPTY_STRING,
+    generatedSecretVisible: false,
+    generatedSecretVersion: 0,
+    settingsOpen: false,
+    settingsClosePending: false,
+    usageExamplesOpen: false,
+    routingSystemPromptOpen: false,
+    providerSystemPromptOpen: false,
+    tenantNameDraft: EMPTY_STRING,
+    tenantRenameDialogOpen: false,
+    tenantNameDirty: false,
+    tenantNameError: EMPTY_STRING,
+    clientKeyReplacementConfirmationOpen: false,
+    clientKeyReplacementPending: false,
+    createTenantDialogOpen: false,
+    createTenantName: EMPTY_STRING,
+    createTenantError: EMPTY_STRING,
+    createTenantPending: false,
+    deleteTenantConfirmationOpen: false,
+    deleteTenantPending: false,
+    discardTenantChangesOpen: false,
+    pendingTenantID: EMPTY_STRING,
+    noticeSurfaces: NOTICE_SURFACES,
+    notice: {
+      kind: NOTICE_KINDS.INFO,
+      message: EMPTY_STRING,
+      surface: NOTICE_SURFACES.HEADER,
+    },
+    noticeDismissTimerID: /** @type {number | null} */ (null),
+    noticeVersion: 0,
+  };
+}
