+++ site/assets/llm-proxy/js/ui/requestExamples.js
@@ -1,0 +1,167 @@
+// @ts-check
+
+import { COPY, NOTICE_KINDS } from "../constants.js?v=20260809i217";
+
+const EMPTY_SECRET_PLACEHOLDER = "<generated-secret>";
+const DEFAULT_TEXT_EXAMPLE_ID = "default-text";
+const DEFAULT_V2_EXAMPLE_ID = "default-v2";
+const DEFAULT_DICTATION_EXAMPLE_ID = "default-dictation";
+const PROVIDER_TEXT_EXAMPLE_ID = "provider-text";
+const PROVIDER_V2_EXAMPLE_ID = "provider-v2";
+const PROVIDER_DICTATION_EXAMPLE_ID = "provider-dictation";
+const JSON_CONTENT_TYPE_HEADER = "Content-Type: application/json";
+const SAMPLE_TEXT_PROMPT = "Hello";
+const SAMPLE_AUDIO_FILE = "recording.webm";
+
+/** @typedef {ReturnType<typeof import("./managementApplicationState.js").createManagementApplicationState>} ManagementApplicationState */
+/** @typedef {ManagementApplicationState & {
+ *   selectedProvider: import("../types.d.js").ProviderProfile | null,
+ *   setSettingsNotice: (kind: string, message: string) => void
+ * }} RequestExampleHost */
+
+/**
+ * @template {object} Responsibility
+ * @param {Responsibility & ThisType<RequestExampleHost & Responsibility>} responsibility
+ * @returns {Responsibility}
+ */
+function requestExampleResponsibility(responsibility) {
+  return responsibility;
+}
+
+/** Create current-profile request-example presentation and clipboard behavior. */
+export function createRequestExamplesResponsibility() {
+  return requestExampleResponsibility({
+    get requestExamples() {
+      const defaultExamples = [];
+      if (this.defaults.provider) {
+        defaultExamples.push(
+          createRequestExample(DEFAULT_TEXT_EXAMPLE_ID, COPY.defaultTextExample, this.defaultTextCurl()),
+          createRequestExample(DEFAULT_V2_EXAMPLE_ID, COPY.defaultV2Example, this.defaultV2Curl()),
+        );
+      }
+      if (this.defaults.dictation_provider) {
+        defaultExamples.push(
+          createRequestExample(DEFAULT_DICTATION_EXAMPLE_ID, COPY.defaultDictationExample, this.defaultDictationCurl()),
+        );
+      }
+      if (!this.selectedProvider) {
+        return defaultExamples;
+      }
+      const providerExamples = [
+        createRequestExample(
+          PROVIDER_TEXT_EXAMPLE_ID,
+          `${this.selectedProvider.label}${COPY.providerTextExampleSuffix}`,
+          this.providerTextCurl(this.selectedProvider),
+        ),
+        createRequestExample(
+          PROVIDER_V2_EXAMPLE_ID,
+          `${this.selectedProvider.label}${COPY.providerV2ExampleSuffix}`,
+          this.providerV2Curl(this.selectedProvider),
+        ),
+      ];
+      if (this.selectedProvider.supports_dictation) {
+        providerExamples.push(
+          createRequestExample(
+            PROVIDER_DICTATION_EXAMPLE_ID,
+            `${this.selectedProvider.label}${COPY.providerDictationExampleSuffix}`,
+            this.providerDictationCurl(this.selectedProvider),
+          ),
+        );
+      }
+      return [...defaultExamples, ...providerExamples];
+    },
+
+    get exampleSecret() {
+      return EMPTY_SECRET_PLACEHOLDER;
+    },
+
+    get proxyOrigin() {
+      return this.runtimeConfig ? this.runtimeConfig.proxyOrigin : window.location.origin;
+    },
+
+    defaultTextCurl() {
+      return [
+        `curl --get ${JSON.stringify(`${this.proxyOrigin}/`)} \\`,
+        `  --data-urlencode 'key=${this.exampleSecret}' \\`,
+        `  --data-urlencode 'prompt=${SAMPLE_TEXT_PROMPT}'`,
+      ].join("\n");
+    },
+
+    defaultV2Curl() {
+      const secret = this.exampleSecret;
+      return [
+        `curl -sS ${JSON.stringify(`${this.proxyOrigin}/v2?key=${secret}`)} \\`,
+        `  -H '${JSON_CONTENT_TYPE_HEADER}' \\`,
+        `  --data '${JSON.stringify({ messages: [{ role: "user", content: SAMPLE_TEXT_PROMPT }] })}'`,
+      ].join("\n");
+    },
+
+    defaultDictationCurl() {
+      return [
+        `curl -sS -X POST ${JSON.stringify(`${this.proxyOrigin}/dictate?key=${this.exampleSecret}`)} \\`,
+        `  -F 'audio=@${SAMPLE_AUDIO_FILE}'`,
+      ].join("\n");
+    },
+
+    /**
+     * @param {import("../types.d.js").ProviderProfile} provider
+     * @returns {string}
+     */
+    providerTextCurl(provider) {
+      return [
+        `curl --get ${JSON.stringify(`${this.proxyOrigin}/`)} \\`,
+        `  --data-urlencode 'key=${this.exampleSecret}' \\`,
+        `  --data-urlencode 'provider=${provider.id}' \\`,
+        `  --data-urlencode 'model=${provider.text_model}' \\`,
+        `  --data-urlencode 'prompt=${SAMPLE_TEXT_PROMPT}'`,
+      ].join("\n");
+    },
+
+    /**
+     * @param {import("../types.d.js").ProviderProfile} provider
+     * @returns {string}
+     */
+    providerV2Curl(provider) {
+      const requestBody = {
+        messages: [{ role: "user", content: SAMPLE_TEXT_PROMPT }],
+        model: provider.text_model,
+      };
+      return [
+        `curl -sS ${JSON.stringify(`${this.proxyOrigin}/v2?key=${this.exampleSecret}&provider=${provider.id}`)} \\`,
+        `  -H '${JSON_CONTENT_TYPE_HEADER}' \\`,
+        `  --data '${JSON.stringify(requestBody)}'`,
+      ].join("\n");
+    },
+
+    /**
+     * @param {import("../types.d.js").ProviderProfile} provider
+     * @returns {string}
+     */
+    providerDictationCurl(provider) {
+      return [
+        `curl -sS -X POST ${JSON.stringify(`${this.proxyOrigin}/dictate?key=${this.exampleSecret}&provider=${provider.id}`)} \\`,
+        `  -F 'audio=@${SAMPLE_AUDIO_FILE}'`,
+      ].join("\n");
+    },
+
+    /** @param {string} command */
+    async copyRequestExample(command) {
+      if (!navigator.clipboard) {
+        this.setSettingsNotice(NOTICE_KINDS.ERROR, COPY.copyUnavailable);
+        return;
+      }
+      await navigator.clipboard.writeText(command);
+      this.setSettingsNotice(NOTICE_KINDS.SUCCESS, COPY.exampleCopied);
+    },
+  });
+}
+
+/**
+ * @param {string} id
+ * @param {string} title
+ * @param {string} command
+ * @returns {import("../types.d.js").RequestExample}
+ */
+function createRequestExample(id, title, command) {
+  return { id, title, command };
+}
