+++ site/assets/llm-proxy/js/ui/runtimeFailure.js
@@ -1,6 +1,6 @@
 // @ts-check
 
-import { COPY, RUNTIME_UI } from "../constants.js?v=20260809b123";
+import { COPY, RUNTIME_UI } from "../constants.js?v=20260809i217";
 
 /**
  * @returns {HTMLElement}
