+++ site/assets/llm-proxy/js/ui/usagePresentation.js
@@ -1,6 +1,6 @@
 // @ts-check
 
-import { DEFAULT_USAGE_INTERVAL, USAGE_INTERVALS } from "../constants.js?v=20260809b123";
+import { DEFAULT_USAGE_INTERVAL, USAGE_INTERVALS } from "../constants.js?v=20260809i217";
 
 const EMPTY_STRING = "";
 const CHART_WIDTH = 640;
