+++ site/resources/copyable-llm-curl-examples/index.html
@@ -174,7 +174,7 @@
             <pre class="usage-snippet"><code>get exampleSecret() {
   return EMPTY_SECRET_PLACEHOLDER;
 }</code></pre>
-            <p>Verified 2026-07-22 by <a href="https://github.com/tyemirov" rel="author">Tyemirov on GitHub</a> against <code>site/assets/llm-proxy/js/ui/keyManagement.js</code>.</p>
+            <p>Verified 2026-07-22 by <a href="https://github.com/tyemirov" rel="author">Tyemirov on GitHub</a> against <code>site/assets/llm-proxy/js/ui/requestExamples.js</code>.</p>
           </section>
 
           <section>
