+++ site/resources/self-service-llm-key-management/index.html
@@ -184,7 +184,7 @@
 if (!this.hasSecret) {
   await this.requestAndApplyGeneratedSecret();
 }</code></pre>
-            <p>Verified 2026-08-08 by <a href="https://github.com/tyemirov" rel="author">Tyemirov on GitHub</a> against <code>site/assets/llm-proxy/js/ui/keyManagement.js</code>.</p>
+            <p>Verified 2026-08-08 by <a href="https://github.com/tyemirov" rel="author">Tyemirov on GitHub</a> against <code>site/assets/llm-proxy/js/ui/authenticationLifecycle.js</code>.</p>
           </section>
 
           <section>
